requirements.txt
aws_cdk.asset_awscli_v1>=2.2.201
aws_cdk.asset_kubectl_v20>=2.1.2
aws_cdk.asset_node_proxy_agent_v6>=2.0.1
aws_cdk_lib>=2.114.1
boto3>=1.33.12
cdk_nag>=2.27.214
constructs>=10.3.0
setuptools>=69.0.2

babel>=2.9.1
certifi>=2023.7.22
configobj>=5.0.6
pyOpenSSL>=24.0.0
cryptography>=42.0.2
fastapi>=0.109.1
jinja2>=3.1.3
oauthlib>=3.2.1
pyjwt>=2.4.0
twisted>=23.10.0rc1
requests>=2.31.0
